split test_plotters.py into multiple files
